operator/requirements.txt
kopf
kubernetes
PyYAML
Requests